+++ cmd/fetch.go
@@ -16,21 +16,25 @@ package cmd
 
 import (
 	"fmt"
+	"strings"
 
+	"github.com/Masterminds/semver"
 	"github.com/spf13/cobra"
 
+	"github.com/tetratelabs/getistio/api"
 	"github.com/tetratelabs/getistio/src/istioctl"
 	"github.com/tetratelabs/getistio/src/manifest"
 	"github.com/tetratelabs/getistio/src/util/logger"
 )
 
+type FetchFlags struct {
+	Name, Version, Flavor string
+	FlavorVersion         int64
+}
+
 func newFetchCmd(homedir string) *cobra.Command {
 	var (
-		flag istioctl.FetchFlags
-		// flagName          string
-		// flagVersion       string
-		// flagFlavor        string
-		// flagFlavorVersion int
+		flag FetchFlags
 	)
 
 	cmd := &cobra.Command{
@@ -80,8 +84,12 @@ For more information, please refer to "getistio list --help" command.
 			if err != nil {
 				return fmt.Errorf("error fetching manifest: %v", err)
 			}
+			d, err := processFetchParams(&flag, ms)
+			if err != nil {
+				return err
+			}
 
-			d, err := istioctl.Fetch(homedir, &flag, ms)
+			d, err = istioctl.Fetch(homedir, d, ms)
 			if err != nil {
 				return err
 			}
@@ -107,4 +115,80 @@ For more information, please refer to "getistio list --help" command.
 	flags.StringVarP(&flag.Flavor, "flavor", "", "", "Flavor of istioctl, e.g. \"--flavor tetrate\" or --flavor tetratefips\" or --flavor istio\". When --name flag is set, this will not be used.")
 	flags.Int64VarP(&flag.FlavorVersion, "flavor-version", "", -1, "Version of the flavor, e.g. \"--version 1\". When --name flag is set, this will not be used.")
 	return cmd
+}
+
+func processFetchParams(flags *FetchFlags,
+	ms *api.Manifest) (*api.IstioDistribution, error) {
+	if len(flags.Name) != 0 {
+		d, err := api.IstioDistributionFromString(flags.Name)
+		if err != nil {
+			return nil, fmt.Errorf("cannot parse given name %s to istio distribution", flags.Name)
+		}
+		return d, nil
+	}
+	if flags.Flavor != api.IstioDistributionFlavorTetrate && flags.Flavor != api.IstioDistributionFlavorTetrateFIPS && flags.Flavor != api.IstioDistributionFlavorIstio {
+		flags.Flavor = api.IstioDistributionFlavorTetrate
+		logger.Infof("fallback to the %s flavor since --flavor flag is not given or not supported\n", flags.Flavor)
+	}
+	if len(flags.Version) == 0 {
+		for _, m := range ms.IstioDistributions {
+			if m.Flavor == flags.Flavor {
+				return m, nil
+			}
+		}
+	}
+
+	ret := &api.IstioDistribution{Version: flags.Version, Flavor: flags.Flavor, FlavorVersion: flags.FlavorVersion}
+
+	if strings.Count(flags.Version, ".") == 1 {
+		// In the case where patch version is not given,
+		// we find the latest patch version
+		var (
+			latest *api.IstioDistribution
+			prev   *semver.Version
+		)
+
+		v, err := semver.NewVersion(flags.Version)
+		if err != nil {
+			return nil, err
+		}
+
+		for _, d := range ms.IstioDistributions {
+			cur, err := semver.NewVersion(d.Version)
+			if err != nil {
+				return nil, err
+			}
+
+			if d.Flavor == ret.Flavor && cur.Minor() == v.Minor() && (prev == nil || cur.GreaterThan(prev)) {
+				prev = cur
+				latest = d
+			}
+		}
+
+		if latest == nil {
+			return nil, fmt.Errorf("invalid version %s", ret.Version)
+		}
+
+		ret.Version = latest.Version
+		logger.Infof("fallback to %s which is the latest patch version in the given verion minor %s\n", ret.Version, flags.Version)
+	}
+
+	if ret.FlavorVersion < 0 {
+		// search the latest flavor version in this flavor
+		var found bool
+		for _, m := range ms.IstioDistributions {
+			if m.Version == ret.Version && m.Flavor == ret.Flavor {
+				ret.FlavorVersion = m.FlavorVersion
+				found = true
+				break
+			}
+		}
+		if !found {
+			return nil, fmt.Errorf("unsupported version=%s and flavor=%s", ret.Version, ret.Flavor)
+		}
+		logger.Infof("fallback to the flavor %d version which is the latest one in %s-%s\n",
+			ret.FlavorVersion, ret.Version, ret.Flavor)
+	}
+
+	return ret, nil
 }
