+++ src/istioctl/istioctl_test.go
@@ -17,7 +17,6 @@ package istioctl
 import (
 	"bytes"
 	"encoding/json"
-	"fmt"
 	"io/ioutil"
 	"os"
 	"os/exec"
@@ -400,152 +399,25 @@ func TestFetch(t *testing.T) {
 	}
 
 	t.Run("not-supported", func(t *testing.T) {
-		for _, c := range []FetchFlags{
+		for _, c := range []*api.IstioDistribution{
 			{Version: "1000.7.4", Flavor: api.IstioDistributionFlavorTetrate},
 			{Version: "1.7.5", Flavor: api.IstioDistributionFlavorTetrateFIPS},
 			{Version: "1.7.5", Flavor: api.IstioDistributionFlavorTetrate, FlavorVersion: 1},
 		} {
-			_, err = Fetch(dir, &c, ms)
+			_, err = Fetch(dir, c, ms)
 			require.Error(t, err)
 		}
 	})
 
 	t.Run("supported", func(t *testing.T) {
-		for _, c := range []FetchFlags{
+		for _, c := range []*api.IstioDistribution{
 			{Version: "1.7.5", Flavor: api.IstioDistributionFlavorTetrate, FlavorVersion: 0},
 			{Version: "1.7.6", Flavor: api.IstioDistributionFlavorTetrate, FlavorVersion: 0},
-			{Name: "1.7.6-tetrate-v0"},
 		} {
-			_, err = Fetch(dir, &c, ms)
+			_, err = Fetch(dir, c, ms)
 			require.NoError(t, err)
 		}
 	})
-}
-
-func Test_processFetchParams(t *testing.T) {
-	type tc struct {
-		flag *FetchFlags
-		mf   *api.Manifest
-		exp  *api.IstioDistribution
-	}
-
-	for i, c := range []tc{
-		{
-			// no args -> fall back to the latest tetrate flavor
-			flag: &FetchFlags{FlavorVersion: -1},
-			mf: &api.Manifest{
-				IstioDistributions: []*api.IstioDistribution{
-					{Version: "1.7.3", FlavorVersion: 0, Flavor: api.IstioDistributionFlavorTetrate},
-				},
-			},
-			exp: &api.IstioDistribution{Version: "1.7.3", FlavorVersion: 0, Flavor: api.IstioDistributionFlavorTetrate},
-		},
-		{
-			// all given
-			flag: &FetchFlags{Version: "1.7.3", FlavorVersion: 100, Flavor: api.IstioDistributionFlavorTetrate},
-			exp:  &api.IstioDistribution{Version: "1.7.3", FlavorVersion: 100, Flavor: api.IstioDistributionFlavorTetrate},
-		},
-		{
-			// given name
-			flag: &FetchFlags{Name: "1.7.3-tetrate-v100"},
-			exp:  &api.IstioDistribution{Version: "1.7.3", FlavorVersion: 100, Flavor: api.IstioDistributionFlavorTetrate},
-		},
-		{
-			// flavor not given
-			flag: &FetchFlags{Version: "1.7.3", FlavorVersion: 100},
-			exp:  &api.IstioDistribution{Version: "1.7.3", FlavorVersion: 100, Flavor: api.IstioDistributionFlavorTetrate},
-		},
-		{
-			//  flavorVersion not given -> fall back to the latest flavor version
-			flag: &FetchFlags{Version: "1.7.3", Flavor: api.IstioDistributionFlavorTetrateFIPS, FlavorVersion: -1},
-			mf: &api.Manifest{
-				IstioDistributions: []*api.IstioDistribution{
-					{Version: "1.7.3", FlavorVersion: 50, Flavor: api.IstioDistributionFlavorTetrateFIPS},
-					{Version: "1.7.3", FlavorVersion: 10000000, Flavor: api.IstioDistributionFlavorTetrate},
-				},
-			},
-			exp: &api.IstioDistribution{Version: "1.7.3", FlavorVersion: 50, Flavor: api.IstioDistributionFlavorTetrateFIPS},
-		},
-		{
-			//  version not given -> choose the latest version given flavor in manifest
-			flag: &FetchFlags{Flavor: api.IstioDistributionFlavorIstio, FlavorVersion: 0},
-			mf: &api.Manifest{
-				IstioDistributions: []*api.IstioDistribution{
-					{Version: "1.7.3", FlavorVersion: 0, Flavor: api.IstioDistributionFlavorTetrateFIPS},
-					{Version: "1.8.3", FlavorVersion: 0, Flavor: api.IstioDistributionFlavorIstio},
-				},
-			},
-			exp: &api.IstioDistribution{Version: "1.8.3", FlavorVersion: 0, Flavor: api.IstioDistributionFlavorIstio},
-		},
-		{
-			//  version and flavor version  not given -> choose the latest version given flavor in manifest
-			flag: &FetchFlags{Flavor: api.IstioDistributionFlavorIstio, FlavorVersion: -1},
-			mf: &api.Manifest{
-				IstioDistributions: []*api.IstioDistribution{
-					{Version: "1.7.3", FlavorVersion: 0, Flavor: api.IstioDistributionFlavorTetrateFIPS},
-					{Version: "1.8.3", FlavorVersion: 0, Flavor: api.IstioDistributionFlavorIstio},
-				},
-			},
-			exp: &api.IstioDistribution{Version: "1.8.3", FlavorVersion: 0, Flavor: api.IstioDistributionFlavorIstio},
-		},
-		{
-			//  flavorVersion not given -> not found error
-			flag: &FetchFlags{Version: "1.7.3", Flavor: api.IstioDistributionFlavorTetrateFIPS, FlavorVersion: -1},
-			mf: &api.Manifest{
-				IstioDistributions: []*api.IstioDistribution{
-					{Version: "1.7.3", FlavorVersion: 50, Flavor: api.IstioDistributionFlavorTetrate},
-				},
-			},
-		},
-		{
-			// flavor, flavorVersion not given -> fall back to the latest tetrate flavor
-			flag: &FetchFlags{Version: "1.7.3", FlavorVersion: -1},
-			mf: &api.Manifest{
-				IstioDistributions: []*api.IstioDistribution{
-					{Version: "1.7.3", FlavorVersion: 100, Flavor: api.IstioDistributionFlavorTetrateFIPS},
-					{Version: "1.7.3", FlavorVersion: 0, Flavor: api.IstioDistributionFlavorTetrate},
-				},
-			},
-			exp: &api.IstioDistribution{Version: "1.7.3", FlavorVersion: 0, Flavor: api.IstioDistributionFlavorTetrate},
-		},
-		{
-			// patch version is not given in 'version', so should fallback to the latest patch version in the minor version
-			flag: &FetchFlags{Version: "1.7", Flavor: api.IstioDistributionFlavorTetrateFIPS, FlavorVersion: -1},
-			mf: &api.Manifest{
-				IstioDistributions: []*api.IstioDistribution{
-					{Version: "1.7.3", FlavorVersion: 100, Flavor: api.IstioDistributionFlavorTetrate},
-					{Version: "1.7.1", FlavorVersion: 100, Flavor: api.IstioDistributionFlavorTetrateFIPS},
-					{Version: "1.8.3", FlavorVersion: 0, Flavor: api.IstioDistributionFlavorTetrate},
-				},
-			},
-			exp: &api.IstioDistribution{Version: "1.7.1", FlavorVersion: 100, Flavor: api.IstioDistributionFlavorTetrateFIPS},
-		},
-		{
-			// patch version is not given in 'version', so should fallback to the latest patch version in the minor version
-			flag: &FetchFlags{Version: "1.7", FlavorVersion: 0},
-			mf: &api.Manifest{
-				IstioDistributions: []*api.IstioDistribution{
-					{Version: "1.7.100", FlavorVersion: 100, Flavor: api.IstioDistributionFlavorTetrate},
-					{Version: "1.7.20", FlavorVersion: 20, Flavor: api.IstioDistributionFlavorTetrate},
-					{Version: "1.7.1", FlavorVersion: 1, Flavor: api.IstioDistributionFlavorTetrate},
-					{Version: "1.7.1", FlavorVersion: 100, Flavor: api.IstioDistributionFlavorTetrateFIPS},
-					{Version: "1.8.3", FlavorVersion: 0, Flavor: api.IstioDistributionFlavorTetrate},
-				},
-			},
-			exp: &api.IstioDistribution{Version: "1.7.100", FlavorVersion: 0, Flavor: api.IstioDistributionFlavorTetrate},
-		},
-	} {
-		t.Run(fmt.Sprintf("%d-th case", i), func(t *testing.T) {
-			actual, err := processFetchParams(c.flag, c.mf)
-			if c.exp == nil {
-				assert.Error(t, err)
-			} else {
-				assert.NoError(t, err)
-				assert.Equal(t, c.exp, actual)
-			}
-		})
-
-	}
 }
 
 func Test_fetch(t *testing.T) {
